style chapter section in homepage
