+++ pages/[lang]/index.tsx
@@ -2,18 +2,23 @@ import { NextPage } from 'next'
 import Link from 'next/link'
 import Grid, { GridDirection } from '@material-ui/core/Grid'
 import { useTheme } from '@material-ui/core/styles'
+import flatten from 'lodash/flatten'
+import { useState, useEffect } from 'react'
+import fetchJsonp from 'fetch-jsonp'
 import useTranslation from '../../hooks/useTranslation'
 import useMediaQuery from '@material-ui/core/useMediaQuery'
 import WithLocale from '../../containers/withLocale'
 import PageLayout from '../../components/PageLayout/PageLayout'
 import LocalSwitcher from '../../components/LocalSwitcher/LocalSwitcher'
 import TextSection from '../../components/Section/TextSection'
-import ChapterSection from '../../components/Section/ChapterSection'
+import ChapterSection, { cities } from '../../components/Section/ChapterSection'
 import Button from '../../components/Button/Button'
 import ContactSection from '../../components/Section/ContactSection'
 import SocialMediaSection from '../../components/Section/SocialMediaSection'
 import TwitterFeed from '../../components/TwitterFeed'
 import Masthead from '../../components/Header/Masthead'
+import Events from '../../components/Events'
+import { mediaquery } from '../../style/style.js'
 
 export const Index: NextPage = () => {
   const { t, locale } = useTranslation()
@@ -21,6 +26,48 @@ export const Index: NextPage = () => {
   const direction: GridDirection = useMediaQuery(theme.breakpoints.up('md'))
     ? 'row'
     : 'column-reverse'
+  const meetupNames = cities
+    .filter(
+      ({ cityContent: { data } }) => !data.is_inactive && data.meetup_name
+    )
+    .map(({ cityContent: { data } }) => data.meetup_name)
+
+  const [events, setEvents] = useState<any>({})
+  const [hasEvents, setHasEvents] = useState(false)
+  const [showMoreLink, setShowMoreLink] = useState(true)
+  const [isLoading, setLoading] = useState(true)
+  useEffect(() => {
+    if (events.firstBatch) {
+      const secondBatch = [...events.allEvents].splice(6, 10)
+      if (!secondBatch.length) setShowMoreLink(false)
+      setEvents({ ...events, secondBatch })
+      return
+    }
+
+    setLoading(true)
+    Promise.all(
+      meetupNames.map(meetupName =>
+        fetchJsonp(`https://api.meetup.com/${meetupName}/events`).then(resp =>
+          resp.json()
+        )
+      )
+    ).then(jsons => {
+      const mixEvents = flatten(jsons.map(({ data }) => data.splice(0, 10)))
+      mixEvents.sort(
+        (a, b) => Date.parse(a.local_date) - Date.parse(b.local_date)
+      )
+
+      if (mixEvents.length) {
+        const firstBatch = [...mixEvents].splice(0, 6)
+        setEvents({ firstBatch, allEvents: mixEvents })
+        setHasEvents(true)
+      } else {
+        setHasEvents(false)
+      }
+    })
+
+    setLoading(false)
+  }, [showMoreLink])
 
   return (
     <PageLayout
@@ -128,7 +175,20 @@ export const Index: NextPage = () => {
         </Grid>
       </TextSection>
 
-      <ChapterSection title={t('chapter.title')} />
+      <TextSection title={t('chapter.title')} anchor='find-events'>
+        <ChapterSection />
+
+        <h4 className='chapter-events'>{t('chapter.events')}</h4>
+        <Events
+          title={t('city.suggestEvent')}
+          events={events}
+          isLoading={isLoading}
+          hasEvents={hasEvents}
+          showMoreLink={showMoreLink}
+          setShowMoreLink={setShowMoreLink}
+          hasMixedGroups
+        />
+      </TextSection>
 
       <TwitterFeed screenName='OpenTechSchool' />
 
@@ -183,6 +243,23 @@ export const Index: NextPage = () => {
         .ots-initiative,
         .ots-community {
           margin-bottom: 30px;
+        }
+
+        .chapter-events {
+          display: none;
+        }
+
+        @media (${mediaquery.tabletToDesktop}) {
+          .chapter-events {
+            display: block;
+            font-family: var(--secondaryFont);
+            font-weight: 500;
+            font-size: 22px;
+            color: #828282;
+            text-align: center;
+            text-transform: uppercase;
+            margin-top: 40px;
+          }
         }
       `}</style>
     </PageLayout>
